+++ opc-ua-sdk/sdk-tests/src/test/java/org/eclipse/milo/opcua/sdk/client/TestNamespace.java
@@ -144,7 +144,7 @@ public class TestNamespace implements Namespace {
             if (node != null) {
                 DataValue value = node.readAttribute(
                     new AttributeContext(context),
-                    id.getAttributeId().intValue(),
+                    id.getAttributeId(),
                     timestamps,
                     id.getIndexRange()
                 );
@@ -176,9 +176,10 @@ public class TestNamespace implements Namespace {
 
                 node.writeAttribute(
                     new AttributeContext(context),
-                    writeValue.getAttributeId().intValue(),
+                    writeValue.getAttributeId(),
                     writeValue.getValue(),
-                    writeValue.getIndexRange());
+                    writeValue.getIndexRange()
+                );
 
                 if (logger.isTraceEnabled()) {
                     Variant variant = writeValue.getValue().getValue();
